% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfmri_sims.R
\name{make_big_case_ctrl_mat}
\alias{make_big_case_ctrl_mat}
\title{make_big_case_ctrl_mat
Makes a big matrix containing all case and control correlation matrices from make_case_ctrl_corrs() function}
\usage{
make_big_case_ctrl_mat(
  case.list,
  ctrl.list,
  save.file = FALSE,
  file.name = NULL,
  rm.lower.tri = F
)
}
\description{
parameters:
}
\details{
case.list - $cases list element of make_case_ctrl_corrs() function (Example: my.corrmats$cases)
ctrl.list - $ctrls list element of make_case_ctrl_corrs() function (Example: my.corrmats$ctrls)
save.file - (logical) will save .rds file if true
file.name - (character) name of file if you would like to save it (do not include .rds extension)

output:

big.case.ctrl.mat - (data.frame) each row is an ROI pair and each column is a subject with case/control identifier
}
