% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsfmri_sims.R
\name{make_case_ctrl_corrs}
\alias{make_case_ctrl_corrs}
\title{make_case_ctrl_corrs}
\usage{
make_case_ctrl_corrs(
  user.adjacency = NULL,
  graph.type = "Erdos-Renyi",
  hi.cor = 0.8,
  lo.cor = -0.8,
  multiway = 2,
  num.variables = 10,
  prob.connected = NULL,
  out.degree = NULL,
  num.case.mats = 1,
  num.ctrl.mats = 1,
  pct.signals = 0.5
)
}
\description{
make_case_ctrl_corrs
}
